fix: System::GetMap returns an std::vector<Eigen::Vector3f>
@fabrizioromanelli nice find: Eigen optimizes dynamic memory allocation/deallocation.
This implies that, if you have both a library and an executable that use such features, they must be
compiled with the same optimization/architectural flags in order to avoid pointers mismatches.
Dynamic memory in Eigen is used for objects with unknown sizes, as Eigen::MatrixXf.
Since it may not be the case for this library (and guess what? It wasn't with the Perceptor), System::GetMap
shall return some more traditional kind of dynamic object.